Use enums instead Propagator impls and simplify DatadogCompositePropagator
